+++ back-end/ChaTop/src/main/java/com/sebastiend/ChaTop/repositories/RentalRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.sebastiend.ChaTop.models.entities.RentalEntity;
 
 @Repository
-public interface RentalRepository extends CrudRepository<RentalEntity, Long> {
+public interface RentalRepository extends CrudRepository<RentalEntity, Integer> {
     
 }
