+++ back-end/ChaTop/src/main/java/com/sebastiend/ChaTop/repositories/UserRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.sebastiend.ChaTop.models.entities.UserEntity;
 
 @Repository
-public interface UserRepository extends CrudRepository<UserEntity, Long> {
+public interface UserRepository extends CrudRepository<UserEntity, Integer> {
     public UserEntity findByEmail(String email);
 }
